Setup Scalafix For CI
Scalafix doesn't support certain rules for certain scala versions. Nevertheless, we'd like to enable these rules. To resolve this, we create a `.scalafix-base.conf` file that is used if the `scalaVersion.value` is not set to the default scala version for the build (2.13.x for now).

`.scalafix.conf` includes `.scalafix-base.conf` and adds in the additional rules.

In addition, this commit enables Scalafix in CI and fixes a handful of Scalafix issues.