Add Code Gen For UnicodeData.txt
This commit adds the code generation code for the data we need to derive from `UnicodeData.txt` (https://www.unicode.org/Public/15.0.0/ucd/UnicodeData.txt).

This includes,

* Code points which are part of the `General_Category=Mark`, e.g. combining marks.
* Code points which have a Virama canonical combining class.
* The bidirectional category for all Unicode code points.

The code generation has been restructured. There is a single entry point for all code generation which ensures that all the code is generated with the same parameters, namely the unicode version.

It should be noted there are not tests in this commit. In the spirit of trying to keep this already rather large commit still reasonably sized, I decided to punt those to another commit. In order to meaningfully test this, we have to not only do the code generation, but also have logic operating on the data for UTS-46, which would make this commit huge.