Implement The Complete Bidirectional Class Mapping
Originally, I had thought that this was covered in the code generated from "UnicodeData.txt", but after experimenting with it I realized that only a subset of code points have their bidirectional mappings present in that file.

This commit adds code generation using the Unicode character data files "PropertyValueAliases.txt" and "DerivedBidiClass.txt" to generate a complete mapping of the bidirectional mapping for each Unicode code point.