Simplify resolveMissingMapping function and difference
This commit simplifies the `resolveMissingMapping` function in the `project` directory used for code generation. It also changes the type of the method `difference` on the `CodePointRange` (also in `project`) from `(Option[CodePointRange], Option[CodePointRange])` to `Option[Ior[CodePointRange, CodePointRange]]`. The two are isomorphic, but the new type is easier to work with.

Also, due to a changes on `main` this commit updates a bunch of header files.