Pass Through Unmodified String On Failure In Mapping Step
Previously when we encountered a disallowed code point in the mapping step, we replaced it with the Unicode replacement character 0xfffd. This is because UTS-46 dictates that we should not render a string with disallowed code points back to the user.

However, the UTS-46 validation steps, which occur after the mapping step, require operating on the input with the disallowed code points in place.

To facilitate this, the `MappingException` has been changed to return both `unsafePartiallyMappedInput` and `renderablePartiallyMappedInput` values. `unsafePartiallyMappedInput` is the mapped input, but disallowed code points are left in place and should never be rendered to the user. `renderablePartiallyMappedInput` is the mapped input, with disallowed code points represented as 0xfffd.